Added DialogueController that is responsible for handling the input after a dialogue is opened.
The new controller makes use of the static Notifier that will tell the UI to update the display as necessary. This will keep the backend logic on what dialogue to display next separate from the UI. This will make it possible to move from different types of dialogue structure without the UI caring.